[CLEANUP] Rename utility classes to service
